requirements.txt
requests==2.31.0
python-dotenv==1.0.1
openai==1.53.0
psycopg[binary]==3.2.3
httpx==0.25.2
requests-toolbelt==1.0.0
Pillow==10.4.0
schedule==1.2.0
streamlit>=1.39.0
replicate==0.15.4
flask==2.3.3
numpy==2.3.4
gunicorn==21.2.0
psutil==5.9.6
python-telegram-bot==20.7
